BLE/microservice/requirements.txt
Flask==2.3.3
bleak==0.21.1
python-dotenv==1.0.0
flask_cors